Upload data loader

+++ structure/skin_lut.py
@@ -1,5 +1,4 @@
 import numpy as np
-from tensorstore import dtype
 
 import utils
 
@@ -15,7 +14,7 @@ class SkinLUT:
             values = np.array(list(skin_lut[i].values()), dtype=float_dtype)
             wavelengths = np.array(list(skin_lut[i].keys()), dtype=float_dtype)
             target_wavelengths = np.linspace(spectrum[0], spectrum[1], spectrum[2])
-            skin_lut[i] = utils.linear_interpolation(values, wavelengths, target_wavelengths, dtype=dtype)
+            skin_lut[i] = utils.linear_interpolation(values, wavelengths, target_wavelengths, dtype=float_dtype)
 
         keys = list(skin_lut.keys())
         data = np.stack([skin_lut[k] for k in keys])
